hacker/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-froala-editor==3.1.0
docopt==0.6.2
idna==2.8
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
pep562==1.0
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.5.2
pymdown-extensions==6.2.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
